+++ praetorian_cli/sdk/test/test_file.py
@@ -12,6 +12,7 @@ class TestFile:
         self.sdk = setup_chariot()
         micro = epoch_micro()
         self.chariot_filepath = f'home/test-file-{micro}.txt'
+        self.encrypted_chariot_filepath = f'_encrypted/test-file-{micro}.txt'
         self.sanitized_filepath = f'home_test-file-{micro}.txt'
         self.bogus_filepath = f'bogus-filepath-{micro}.txt'
         self.local_filepath = f'./test-file-{micro}.txt'
@@ -51,6 +52,21 @@ class TestFile:
         with pytest.raises(Exception) as ex_info:
             self.sdk.files.get(self.chariot_filepath)
         assert str(ex_info.value) == f'File {self.chariot_filepath} not found.'
+
+    def test_add_encrypted_file(self):
+        self.sdk.files.add(self.local_filepath, self.encrypted_chariot_filepath)
+        files, offset = self.sdk.files.list(self.encrypted_chariot_filepath)
+        assert files[0]['name'] == self.encrypted_chariot_filepath
+
+    def test_get_encrypted_file(self):
+        content = self.sdk.files.get_utf8(self.encrypted_chariot_filepath)
+        assert content == self.content
+
+    def test_delete_encrypted_file(self):
+        self.sdk.files.delete(self.encrypted_chariot_filepath)
+        with pytest.raises(Exception) as ex_info:
+            self.sdk.files.get(self.encrypted_chariot_filepath)
+        assert str(ex_info.value) == f'File {self.encrypted_chariot_filepath} not found.'
 
     def teardown_class(self):
         os.remove(self.local_filepath)