+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# NEXT_RELEASE
+- [New Feature] Added support for the _encrypted/ encrypted file store partition.
+
 # 2.2.12
 - [Bug fix] Updated the risk's affected assets query to handle ports and webpages.